编写"报警巡查"后端实时监控类文件及报警信息持久化功能代码

+++ src/main/java/com/ktcn/entity/Alarmpatrol.java
@@ -17,8 +17,8 @@ public class Alarmpatrol {
 	private String ap_name;			// 点位名
 	private String describe;		// 描述
 	private String measured_value;	// 实测值
-	private float upDiff;			// 上限差值(上限值减实测值,负值为异常)
-	private float lowDiff;			// 下限差值(下限值减实测值,负值为异常)
+	private double upDiff;			// 上限差值(上限值减实测值,负值为异常)
+	private double lowDiff;			// 下限差值(下限值减实测值,负值为异常)
 	@JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
 	private Date ap_time;			// 报警时间
 	private int ap_state;			// 状态码
